refactor: reduce mailbox login change scope
