+++ CHANGELOG.md
@@ -37,7 +37,6 @@
 
 ### Testing
 
-- test: |邮箱登录| 覆盖密码登录 JWT 有效期与续期、非法 token 类型，以及用户重置绑定邮箱密码的权限检查
 - test: |E2E| 覆盖 D1 数据库大小响应、配置键隔离，以及数据库页面套餐选择的持久化与刷新恢复
 - fix: |E2E| 覆盖发信页面草稿编辑、正文格式切换及 HTML 预览
 - fix: |E2E| 覆盖用户 JWT 发信接口的地址归属、额度扣减、实际投递和发件箱操作，以及用户中心查看地址凭证、切换发件地址和按地址过滤发件箱的完整流程
