+++ vitepress-docs/docs/en/guide/feature/admin-user-management.md
@@ -1,9 +1,5 @@
 # Admin User Management
 
-::: info Mailbox password login
-See [mailbox password login](./mailbox-password-login) for password-only login, login JWT renewal and bound-mailbox password resets.
-:::
-
 ## User Management Page
 
 ![admin-user-management](/feature/admin-user-management.png)
