+++ vitepress-docs/docs/zh/guide/feature/admin-user-management.md
@@ -1,9 +1,5 @@
 # Admin 用户相关
 
-::: info 邮箱密码登录
-仅密码登录、登录 JWT 自动续期及绑定邮箱密码重置见[邮箱密码登录](./mailbox-password-login)。
-:::
-
 ## 用户管理页面
 
 ![admin-user-management](/feature/admin-user-management.png)
